Small API changes
